start moving generated structs to their own module
